requirements.txt
# Core Backend API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7

# Graph & Data Structures
networkx>=3.2.1
gremlinpython>=3.7.0
matplotlib>=3.8.2

# Source Code Parsing
tree-sitter==0.21.3
tree-sitter-languages>=1.10.0

# Version Control
gitpython>=3.1.41

# Utilities
requests>=2.31.0
pydantic>=2.6.0
python-dotenv>=1.0.0
boto3>=1.34.0

# AI & Vector Search
chromadb>=0.4.22
opensearch-py>=2.4.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.9
langchain-aws>=0.1.0
sentence-transformers>=2.3.1